requirements.txt
coverage==7.6.1
flake8==7.1.1
requests==2.32.3
beautifulsoup4==4.12.3
iteration_utilities==0.12.1

# Airflow providers (http://airflow.apache.org/docs/apache-airflow-providers/packages-ref.html)
apache-airflow-providers-mysql==5.6.3
